OSIS-2805 Use CustomQueryset for annotation
